#2239 move existing tests for collection iterator to a separate class
+ simplify them: use a separate html page with only one select.
its values/labels are not easily readable.
